Upload three preprocessing methods
This file contains three preprocessing methods
(1) Count the number of waiting instance, ie the same latitude and longitude tuple that has change as the time progresses. 
(2) Destination counting for each nearest stand. This should be a visualization and good reason to why we choose to associate taxi stand to the starting point of the trip to better predict the trajectory.
(3) Convert timestamp in Unix time to hours.